Somewhat solve Task 7 and 8

+++ ComputationalPhysics/E1/Task7/func.h
@@ -8,7 +8,7 @@ Further developed by Martin Gren on 2015-10-23.
 #ifndef _E1_func_h
 #define _E1_func_h
 
-extern void calc_acc(double *, double *, double, double, int);
+extern void calc_acc(double *, double *, double, double, double);
 
 extern double calc_pe(double *, double, int);
